+++ src/main/java/com/group7/serviceInterface/BaseDataServiceLocal.java
@@ -40,7 +40,7 @@ public interface BaseDataServiceLocal {
 	public void putEventCauseData(Collection<EventCause> eventCauseList);
 	public void putFailureData(Collection<Failure>failureList);
 
-	public Collection<Long> getTotalFailuresOfSpecificPhone(BigInteger phoneType, String startDate, String endDate);
+	public Collection<Long> getTotalFailuresOfSpecificPhone(int phoneType, String startDate, String endDate);
 	public Collection<Long> getTotalFailuresOfSpecificIMSI(BigInteger imsi, String startDate, String endDate);
 	public Collection<Object> getAllCallFailuresAndTotalDurationPerIMSI(BigInteger imsi, String startDate, String endDate);
 	public Collection<Object> getAllUniqueEventCausecodeCombinations(String model);
